Prompt for task deletion. Fixed adapter bug and jittering.

+++ app/src/main/java/com/aechkae/organizer/focus/adapters/ActiveTaskRVAdapter.java
@@ -1,6 +1,10 @@
 package com.aechkae.organizer.focus.adapters;
 
+import android.content.Context;
+import android.content.DialogInterface;
+import android.support.v7.app.AlertDialog;
 import android.support.v7.widget.RecyclerView;
+import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -8,7 +12,7 @@ import android.view.ViewGroup;
 import com.aechkae.organizer.database.OrgDBAdapter;
 import com.aechkae.organizer.databinding.ActiveTaskItemBinding;
 import com.aechkae.organizer.focus.UncompTask;
-import com.aechkae.organizer.focus.Task;
+import com.aechkae.organizer.focus.Task.TaskType;
 import com.aechkae.organizer.R;
 import com.aechkae.organizer.focus.view_holders.ActiveTaskViewHolder;
 
@@ -18,12 +22,15 @@ import static com.aechkae.organizer.focus.Task.UNCOMP_DESC_CUTOFF_LENGTH;
 
 public class ActiveTaskRVAdapter extends RecyclerView.Adapter<ActiveTaskViewHolder>{
     private OrgDBAdapter db_adapter;
-
+    private Context mContext;
     private List<UncompTask> activeTasks;
 
-    public ActiveTaskRVAdapter(List<UncompTask> tasks, OrgDBAdapter db_adapter){
+    String TAG = "t";
+
+    public ActiveTaskRVAdapter(List<UncompTask> tasks, OrgDBAdapter db_adapter, Context context){
         activeTasks = tasks;
         this.db_adapter = db_adapter;
+        mContext = context;
     }
 
     public void setActiveTasks(List<UncompTask> activeTasks) {
@@ -51,31 +58,42 @@ public class ActiveTaskRVAdapter extends RecyclerView.Adapter<ActiveTaskViewHold
 
 
         //Remove task from DB
-        holder.getActiveTaskItemBinding().activeTaskDeleteBtn.setOnClickListener((view) -> {
-            db_adapter.deleteTask(
-                    new String[]{vhBinding.activeTaskCode.getText().toString()});
-            activeTasks.remove(position);
-            notifyItemChanged(position);
+        vhBinding.activeTaskDeleteBtn.setOnClickListener((view) -> {
+            AlertDialog.Builder builder = new AlertDialog.Builder(mContext);
+            builder.setMessage(R.string.focus_delete_task_prompt);
+            builder.setCancelable(false);
+            builder.setNegativeButton(R.string.cancel_action, (dialog, id) -> {});
+            builder.setPositiveButton(R.string.confirm_deletion, (dialog, id) -> {
+                //TODO update
+                db_adapter.deleteTask(
+                        new String[]{vhBinding.activeTaskCode.getText().toString()});
+                activeTasks.remove(holder.getAdapterPosition());
+                notifyItemRemoved(holder.getAdapterPosition());
+            });
+
+            AlertDialog alert = builder.create();
+            alert.show();
         });
 
 
         //Move to backlog
-        holder.getActiveTaskItemBinding().activeMvBacklogBtn.setOnClickListener((view) -> {
+        vhBinding.activeMvBacklogBtn.setOnClickListener((view) -> {
+            Log.d(TAG, "pre-move");
             db_adapter.moveTask(
-                    new String[]{vhBinding.activeTaskCode.getText().toString()}, Task.TaskType.BACKLOG);
-            activeTasks.remove(position);
-            notifyItemChanged(position);
+                    new String[]{vhBinding.activeTaskCode.getText().toString()}, TaskType.BACKLOG);
+            activeTasks.remove(holder.getAdapterPosition());
+            notifyItemRemoved(holder.getAdapterPosition());
         });
 
         //Complete task
-        holder.getActiveTaskItemBinding().completeTaskBtn.setOnClickListener((view) -> {
+        vhBinding.completeTaskBtn.setOnClickListener((view) -> {
             db_adapter.completeTask(
                     vhBinding.activeTaskCode.getText().toString(),
                     vhBinding.activeTaskDesc.getText().toString()
             );
 
-            activeTasks.remove(position);
-            notifyItemChanged(position);
+            activeTasks.remove(holder.getAdapterPosition());
+            notifyItemRemoved(holder.getAdapterPosition());
         });
     }
