+++ app/src/main/java/com/aechkae/organizer/focus/adapters/BacklogTaskRVAdapter.java
@@ -1,14 +1,15 @@
 package com.aechkae.organizer.focus.adapters;
 
+import android.content.Context;
+import android.support.v7.app.AlertDialog;
 import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
 import com.aechkae.organizer.database.OrgDBAdapter;
-import com.aechkae.organizer.databinding.ActiveTaskItemBinding;
 import com.aechkae.organizer.databinding.BacklogTaskItemBinding;
-import com.aechkae.organizer.focus.Task;
+import com.aechkae.organizer.focus.Task.TaskType;
 import com.aechkae.organizer.R;
 import com.aechkae.organizer.focus.UncompTask;
 import com.aechkae.organizer.focus.view_holders.BacklogTaskViewHolder;
@@ -19,11 +20,13 @@ import static com.aechkae.organizer.focus.Task.UNCOMP_DESC_CUTOFF_LENGTH;
 
 public class BacklogTaskRVAdapter extends RecyclerView.Adapter<BacklogTaskViewHolder>{
     private OrgDBAdapter db_adapter;
+    private Context mContext;
     private List<UncompTask> backlogTasks;
 
-    public BacklogTaskRVAdapter(List<UncompTask> tasks, OrgDBAdapter dbAdapter){
+    public BacklogTaskRVAdapter(List<UncompTask> tasks, OrgDBAdapter dbAdapter, Context context){
         db_adapter = dbAdapter;
         backlogTasks = tasks;
+        mContext = context;
     }
 
     @Override
@@ -50,20 +53,28 @@ public class BacklogTaskRVAdapter extends RecyclerView.Adapter<BacklogTaskViewHo
         vhBinding.taskPercDone.setText(vhTask.perc_done);
 
         //Remove task from DB
-        holder.getBacklogTaskBinding().backlogDeleteBtn.setOnClickListener((view) -> {
-            db_adapter.deleteTask(
-                    new String[]{holder.getBacklogTaskBinding().taskCode.getText().toString()});
-            backlogTasks.remove(position);
-            notifyItemChanged(position);
+        vhBinding.backlogDeleteBtn.setOnClickListener((view) -> {
+            AlertDialog.Builder builder = new AlertDialog.Builder(mContext);
+            builder.setMessage(R.string.focus_delete_task_prompt);
+            builder.setCancelable(false);
+            builder.setNegativeButton(R.string.cancel_action, (dialog, id) -> {});
+            builder.setPositiveButton(R.string.confirm_deletion, (dialog, id) -> {
+                db_adapter.deleteTask(
+                        new String[]{vhBinding.taskCode.getText().toString()});
+                backlogTasks.remove(holder.getAdapterPosition());
+                notifyItemRemoved(holder.getAdapterPosition());
+            });
+
+            AlertDialog alert = builder.create();
+            alert.show();
         });
 
         //Move to active
-        holder.getBacklogTaskBinding().backlogMvActiveBtn.setOnClickListener((view) -> {
+        vhBinding.backlogMvActiveBtn.setOnClickListener((view) -> {
             db_adapter.moveTask(
-                    new String[]{holder.getBacklogTaskBinding().
-                            taskCode.getText().toString()}, Task.TaskType.ACTIVE);
-            backlogTasks.remove(position);
-            notifyItemChanged(position);
+                    new String[]{vhBinding.taskCode.getText().toString()}, TaskType.ACTIVE);
+            backlogTasks.remove(holder.getAdapterPosition());
+            notifyItemRemoved(holder.getAdapterPosition());
         });
     }
